+++ playgrounds/skeleton-svelte/src/routes/components/popover/+page.svelte
@@ -38,13 +38,13 @@
 			triggerBase="btn preset-tonal"
 			contentBase="card bg-surface-200-800 p-4 space-y-4 max-w-[320px]"
 			arrow
-			arrowBackground="!bg-surface-200 dark:!bg-surface-800"
+			arrowBackground="var(--color-surface-200-800)"
 		>
 			{#snippet trigger()}Click Me{/snippet}
 			{#snippet content()}
 				<header class="flex justify-between">
 					<p class="font-bold text-xl">Popover</p>
-					<button class="btn-icon" onclick={popoverClose}>
+					<button class="btn-icon" onclick={popoverClose} title="Close" aria-label="Close">
 						<IconX />
 					</button>
 				</header>